at32f4: timer: split out common part
