+++ lib/at32/common/timer_common_at32.c
